CascadeData: Experimental Data of Cascade Experiments in Genomics

These experimental expression data (5 leukemic 'CLL' B-lymphocyte of aggressive form from 'GSE39411', doi:10.1073/pnas.1211130110 ), after B-cell receptor stimulation, are used as examples by packages such as the 'Cascade' one, a modeling tool allowing gene selection, reverse engineering, and prediction in cascade networks. Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2014) doi:10.1093/bioinformatics/btt705 .

References

Jung, N., Bertrand, F., Bahram, S., Vallat, L., and Maumy-Bertrand, M. (2013). Cascade: a R-package to study, predict and simulate the diffusion of a signal through a temporal gene network. Bioinformatics, btt705.

Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. Proceedings of the National Academy of Sciences, 110(2), 459-464.

See also

Author

This package has been written by Frederic Bertrand, Myriam Maumy-Bertrand and Nicolas Jung with biological insights from Laurent Vallat.

Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>